Take the opportunity to refactor the widgets packaging...
* split the huge package.scala in widgets in several files, keeping the package object for backward-compat (using out directly f.i.)
* fixed the cleaning of the Geo widget when circles were used and applyOn was called (cc @velvia)
* fixed implicit conversion to ToPoints for List and so on (cc @vidma, see https://github.com/andypetrella/spark-notebook/commit/752227eb73123e04dc372df73ba871d97278e624#commitcomment-16855319 :-))
* added a reactive html widget that would use string to define whatever html we may want to render reactively (can be changed using apply)
